Get a lot more picky with our compiler warnings.

+++ src/include/efibootmgr.h
@@ -40,8 +40,8 @@ typedef struct {
 	int bootnext;
 	int verbose;
 	int active;
-	uint32_t acpi_hid;
-	uint32_t acpi_uid;
+	int64_t acpi_hid;
+	int64_t acpi_uid;
 	unsigned int delete_boot:1;
 	unsigned int delete_bootorder:1;
 	unsigned int delete_bootnext:1;
